requirements.txt
lxml
Pillow
XlsxWriter
python-pptx